data_research_agent/requirements.txt
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pandas>=2.2.0
pydantic>=2.6.0
pyyaml>=6.0.0
apscheduler>=3.10.0
python-dateutil>=2.9.0
pytest>=8.0.0
